requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
imbalanced-learn
pyyaml
joblib